backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
SQLAlchemy==2.0.48
python-dotenv==1.0.0
